+++ modules/video_output/x11/xcommon.h
@@ -242,11 +242,11 @@ struct picture_sys_t
 
 typedef struct mwmhints_t
 {
-    uint32_t flags;
-    uint32_t functions;
-    uint32_t decorations;
-    int32_t  input_mode;
-    uint32_t status;
+    unsigned long flags;
+    unsigned long functions;
+    unsigned long decorations;
+    signed   long input_mode;
+    unsigned long status;
 } mwmhints_t;
 
 /*****************************************************************************
